# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CMPLR API Documentation",
        "description": "Welcome to the CMPLR API Documentation!",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "CMPLR API Server"
        }
    ],
    "paths": {
...
        "/{blog-identifier}/notes/delete": {
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "unfollow tag",
                "operationId": "unfollow tag",
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "note_tumblelog",
                        "in": "query",
                        "description": "the name of blog containing post",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Meta": {
                                            "properties": {
                                                "Status": {
                                                    "type": "integer",
                                                    "example": 200
                                                },
                                                "msg": {
                                                    "type": "string",
                                                    "example": "ok"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "user/like": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Like a Post",
                "description": "enables the user to like a post through the post id",
                "operationId": "UserLike",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "The ID of the post to like",
                        "required": true,
                        "schema": {
                            "type": "Number"
                        }
                    },
                    {
                        "name": "reblog_key",
                        "in": "query",
                        "description": "The reblog key for the post id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id",
                                    "reblog_key"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": 1
                                    },
                                    "reblog_key": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "hello123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "200": {
                        "description": "success"
                    }
                }
            }
        },
        "user/unlike": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Unlike a Post",
                "description": "enables the user to unlike a post through the post id",
                "operationId": "UserUnlike",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "The ID of the post to unlike",
                        "required": true,
                        "schema": {
                            "type": "Number"
                        }
                    },
                    {
                        "name": "reblog_key",
                        "in": "query",
                        "description": "The reblog key for the post id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id",
                                    "reblog_key"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": 1
                                    },
                                    "reblog_key": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "hello123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "200": {
                        "description": "success"
                    }
                }
            }
        },
        "user/post/reply": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "add new reply to Specific Post",
                "operationId": "add reply",
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "reblog_key",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "reply_text",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "text"
                        }
                    },
                    {
                        "name": "tumblelog",
                        "in": "query",
                        "description": "the name of blog containing post to get if he is the original poster",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "201": {
                        "description": "created Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Meta": {
                                            "properties": {
                                                "Status": {
                                                    "type": "integer",
                                                    "example": 200
                                                },
                                                "msg": {
                                                    "type": "string",
                                                    "example": "ok"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "user/tags/add": {
            "post": {
                "tags": [
                    "Tags"
                ],
                "summary": "follow new tag",
                "operationId": "follow tag",
                "parameters": [
                    {
                        "name": "tag_name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Meta": {
                                            "properties": {
                                                "Status": {
                                                    "type": "integer",
                                                    "example": 200
                                                },
                                                "msg": {
                                                    "type": "string",
                                                    "example": "ok"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "user/tags/remove": {
            "delete": {
                "tags": [
                    "Tags"
                ],
                "summary": "unfollow tag",
                "operationId": "unfollow tag",
                "parameters": [
                    {
                        "name": "tag_name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Meta": {
                                            "properties": {
                                                "Status": {
                                                    "type": "integer",
                                                    "example": 200
                                                },
                                                "msg": {
                                                    "type": "string",
                                                    "example": "ok"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}